# Cache fourier sin/cos computation