Performance comparison for one event:

Here:
```
❯ ipython time_fourier.py
[[-0.84497639  0.12606332 -0.41890738 ... -0.04808525 -0.29396779
  -0.9354359 ]
 [ 0.23389281 -0.00328971  1.44855959 ...  0.24069143 -2.06059856
  -0.30701159]]
439 µs ± 693 ns per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```

On main:
```
❯ ipython time_fourier.py
[[-0.84497639  0.12606332 -0.41890738 ... -0.04808525 -0.29396779
  -0.9354359 ]
 [ 0.23389281 -0.00328971  1.44855959 ...  0.24069143 -2.06059856
  -0.30701159]]
940 µs ± 5.6 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```


Script:
```
from ctapipe_io_lst import LSTEventSource
from pathlib import Path
from traitlets.config import Config


test_data = Path('test_data').absolute()
test_r0_path = test_data / 'real/R0/20200218/LST-1.1.Run02008.0000_first50.fits.fz'
calib_version = "ctapipe-v0.17"
calib_path = test_data / 'real/monitoring/PixelCalibration/Cat-A/'
test_time_calib_path = calib_path / f'drs4_time_sampling_from_FF/20191124/{calib_version}/time_calibration.Run01625.0000.h5'
test_calib_path = calib_path / f'calibration/20200218/{calib_version}/calibration_filters_52.Run02006.0000.h5'
test_drs4_pedestal_path = calib_path / f'drs4_baseline/20200218/{calib_version}/drs4_pedestal.Run02005.0000.h5'


config = Config({
    'LSTEventSource': {
        'pointing_information': False,
        'trigger_information': False,
        'LSTR0Corrections': {
            'drs4_pedestal_path': test_drs4_pedestal_path,
            'drs4_time_calibration_path': test_time_calib_path,
            'calibration_path': test_calib_path,
        },
    },
})

source = LSTEventSource(
    input_url=test_r0_path,
    config=config,
)

with source:
    event = next(iter(source))

tel_id = 1


def compute_timeshift():
    return source.r0_r1_calibrator.get_drs4_time_correction(
        tel_id, source.r0_r1_calibrator.first_cap[tel_id],
        selected_gain_channel=event.r1.tel[tel_id].selected_gain_channel,
    )

print(compute_timeshift())

result = get_ipython().run_line_magic("timeit", "compute_timeshift()")
# orint"mean": result.average, "std": result.stdev})
```

Run with `ipython time_fourier.py`